refactor(scripts): share update manifest logic and fix build arch detection
